made animation responsive to scroll
